requirements.txt
tornado
salt
apache-libcloud
python-novaclient
pbkdf2
